+++ test_bytes2.py
@@ -1,4 +1,0 @@
-from webauthn.helpers import bytes_to_base64url
-
-b = b"\x87\x00"
-print(bytes_to_base64url(b))